requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
exceptiongroup==1.2.2
flet==0.25.1
flet-desktop==0.25.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
MouseInfo==0.1.3
oauthlib==3.2.2
pillow==11.0.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==10.3.2
pyobjc-framework-Cocoa==10.3.2
pyobjc-framework-Quartz==10.3.2
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
PyYAML==6.0.2
repath==0.9.0
rubicon-objc==0.4.9
six==1.16.0
sniffio==1.3.1
typing_extensions==4.12.2
